refactor: 클래스 및 함수 이름 변
