+++ app/src/main/java/woowacourse/payments/ui/addcard/AddCardTopBar.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.newcard
+package woowacourse.payments.ui.addcard
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
@@ -16,18 +16,18 @@ import woowacourse.payments.R
 @Suppress("ktlint:standard:function-naming")
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun NewCardTopBar(
+fun AddCardTopBar(
     onBackClick: () -> Unit,
     onSaveClick: () -> Unit,
     modifier: Modifier = Modifier,
 ) {
     TopAppBar(
-        title = { Text(stringResource(R.string.new_card_title)) },
+        title = { Text(stringResource(R.string.add_card_title)) },
         navigationIcon = {
             IconButton(onClick = { onBackClick() }) {
                 Icon(
                     imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                    contentDescription = stringResource(R.string.new_card_back_button_description),
+                    contentDescription = stringResource(R.string.add_card_back_button_description),
                 )
             }
         },
@@ -35,7 +35,7 @@ fun NewCardTopBar(
             IconButton(onClick = { onSaveClick() }) {
                 Icon(
                     imageVector = Icons.Filled.Check,
-                    contentDescription = stringResource(R.string.new_card_save_button_description),
+                    contentDescription = stringResource(R.string.add_card_save_button_description),
                 )
             }
         },
