+++ app/src/main/java/woowacourse/payments/ui/addcard/textfields/PasscodeTextField.kt
@@ -1,25 +1,25 @@
-package woowacourse.payments.ui.newcard.textfields
+package woowacourse.payments.ui.addcard.textfields
 
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.focus.FocusDirection
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.PasswordVisualTransformation
 import woowacourse.payments.R
-import woowacourse.payments.domain.CardNumber
-import woowacourse.payments.ui.formatter.CardNumberFormat
+import woowacourse.payments.domain.Passcode
 import woowacourse.payments.ui.theme.Gray
+
+private const val PASSCODE_REQUIRED_LENGTH = 4
 
 @Suppress("ktlint:standard:function-naming")
 @Composable
-fun CardNumberTextField(
+fun PasscodeTextField(
     text: MutableState<String>,
     isError: MutableState<Boolean>,
 ) {
@@ -27,40 +27,39 @@ fun CardNumberTextField(
 
     fun updateValue(newValue: String) {
         val filteredValue: String =
-            newValue.filter(Char::isDigit).take(CardNumberFormat.REQUIRED_LENGTH)
+            newValue.filter(Char::isDigit).take(PASSCODE_REQUIRED_LENGTH)
 
         text.value = filteredValue
-        isError.value = runCatching { CardNumber(text.value) }.isFailure
+        isError.value = runCatching { Passcode(text.value) }.isFailure
 
-        if (!isError.value && filteredValue.length == CardNumberFormat.REQUIRED_LENGTH) {
-            focusManager.moveFocus(FocusDirection.Next)
+        if (!isError.value && filteredValue.length == PASSCODE_REQUIRED_LENGTH) {
+            focusManager.clearFocus()
         }
     }
 
     OutlinedTextField(
-        modifier = Modifier.fillMaxWidth(),
+        modifier = Modifier.fillMaxWidth(0.5F),
         value = text.value,
         onValueChange = { newValue: String -> updateValue(newValue) },
         singleLine = true,
-        visualTransformation = CardNumberFormat.visualTransformation,
-        label = { Text(stringResource(R.string.card_number_label)) },
+        visualTransformation = PasswordVisualTransformation(),
+        label = { Text(stringResource(R.string.passcode_label)) },
         placeholder = {
             Text(
-                text = stringResource(R.string.card_number_placeholder),
+                text = stringResource(R.string.passcode_placeholder),
                 color = Gray,
             )
         },
         supportingText = {
             Text(
                 if (isError.value) {
-                    stringResource(R.string.card_number_error_message)
+                    stringResource(R.string.passcode_error_message)
                 } else {
                     ""
                 },
             )
         },
         isError = isError.value,
-        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
-        keyboardActions = KeyboardActions(onDone = { focusManager.moveFocus(FocusDirection.Next) }),
+        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.NumberPassword),
     )
 }
